Texture mapping done with no alpha

+++ site/src/client/render/shaders/basic/Vertex.ts
@@ -9,7 +9,7 @@ export const vsSource = `
     uniform mat4 uModelViewMatrix;
     uniform mat4 uProjectionMatrix;
 
-    varying highp vec2 vTextureCoord;
+    varying lowp vec2 vTextureCoord;
 
     void main() {
       mat4 billboardMatrix = mat4(
